requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
requests>=2.28.0
django-bootstrap4>=22.1
newsapi-python>=0.2.7
gunicorn>=20.1.0
whitenoise>=6.0.0
python-dotenv>=0.19.0 